+++ src/app/user/user.module.js
@@ -3,6 +3,6 @@
   // 用户模块, 负责用户状态记录, 业务执行
 
   angular
-    .module('pf.user', ['pf.core']);
+    .module('pf.user', ['pf.core', 'pf.condition']);
 
 })();